+++ app/use_cases/petition_plugin/detail_updater.rb
@@ -31,6 +31,7 @@ module PetitionPlugin
         :call_to_action,
         :initial_signatures_goal,
         :signatures_required,
+        :requires_mobile_validation,
         :presentation,
         :video_id,
         :scope_coverage,