+++ db/migrate/20191121163237_add_requires_mobile_validation_to_petition_details.rb
@@ -1,0 +1,5 @@
+class AddRequiresMobileValidationToPetitionDetails < ActiveRecord::Migration
+  def change
+    add_column :petition_plugin_details, :requires_mobile_validation, :boolean, default: false, null: false
+  end
+end